Enum swc_css_ast::AtRulePrelude

pub enum AtRulePrelude {

    ListOfComponentValues(ListOfComponentValues),
    CharsetPrelude(Str),
    PropertyPrelude(CustomPropertyName),
    CounterStylePrelude(CustomIdent),
    ColorProfilePrelude(ColorProfileName),
    DocumentPrelude(DocumentPrelude),
    FontPaletteValuesPrelude(DashedIdent),
    FontFeatureValuesPrelude(FontFeatureValuesPrelude),
    NestPrelude(SelectorList),
    KeyframesPrelude(KeyframesName),
    ImportPrelude(ImportPrelude),
    NamespacePrelude(NamespacePrelude),
    MediaPrelude(MediaQueryList),
    SupportsPrelude(SupportsCondition),
    PagePrelude(PageSelectorList),
    LayerPrelude(LayerPrelude),
    ContainerPrelude(ContainerCondition),
    CustomMediaPrelude(CustomMediaQuery),
    ScopePrelude(ScopeRange),

}

Variants

ListOfComponentValues(ListOfComponentValues)

CharsetPrelude(Str)

PropertyPrelude(CustomPropertyName)

CounterStylePrelude(CustomIdent)

ColorProfilePrelude(ColorProfileName)

DocumentPrelude(DocumentPrelude)

FontPaletteValuesPrelude(DashedIdent)

FontFeatureValuesPrelude(FontFeatureValuesPrelude)

NestPrelude(SelectorList)

KeyframesPrelude(KeyframesName)

ImportPrelude(ImportPrelude)

NamespacePrelude(NamespacePrelude)

MediaPrelude(MediaQueryList)

SupportsPrelude(SupportsCondition)

PagePrelude(PageSelectorList)

LayerPrelude(LayerPrelude)

ContainerPrelude(ContainerCondition)

CustomMediaPrelude(CustomMediaQuery)

ScopePrelude(ScopeRange)

Implementations

impl AtRulePrelude

pub const fn is_list_of_component_values(&self) -> bool

Returns true if self is of variant ListOfComponentValues.

pub fn as_list_of_component_values(&self) -> Option<&ListOfComponentValues>

Returns Some if self is a reference of variant ListOfComponentValues, and None otherwise.

pub fn as_mut_list_of_component_values( &mut self ) -> Option<&mut ListOfComponentValues>

Returns Some if self is a mutable reference of variant ListOfComponentValues, and None otherwise.

pub fn expect_list_of_component_values(self) -> ListOfComponentValues
where Self: Debug,

Unwraps the value, yielding the content of ListOfComponentValues.

Panics

Panics if the value is not ListOfComponentValues, with a panic message including the content of self.

pub fn list_of_component_values(self) -> Option<ListOfComponentValues>

Returns Some if self is of variant ListOfComponentValues, and None otherwise.

pub const fn is_charset_prelude(&self) -> bool

Returns true if self is of variant CharsetPrelude.

pub fn as_charset_prelude(&self) -> Option<&Str>

Returns Some if self is a reference of variant CharsetPrelude, and None otherwise.

pub fn as_mut_charset_prelude(&mut self) -> Option<&mut Str>

Returns Some if self is a mutable reference of variant CharsetPrelude, and None otherwise.

pub fn expect_charset_prelude(self) -> Str
where Self: Debug,

Unwraps the value, yielding the content of CharsetPrelude.

Panics

Panics if the value is not CharsetPrelude, with a panic message including the content of self.

pub fn charset_prelude(self) -> Option<Str>

Returns Some if self is of variant CharsetPrelude, and None otherwise.

pub const fn is_property_prelude(&self) -> bool

Returns true if self is of variant PropertyPrelude.

pub fn as_property_prelude(&self) -> Option<&CustomPropertyName>

Returns Some if self is a reference of variant PropertyPrelude, and None otherwise.

pub fn as_mut_property_prelude(&mut self) -> Option<&mut CustomPropertyName>

Returns Some if self is a mutable reference of variant PropertyPrelude, and None otherwise.

pub fn expect_property_prelude(self) -> CustomPropertyName
where Self: Debug,

Unwraps the value, yielding the content of PropertyPrelude.

Panics

Panics if the value is not PropertyPrelude, with a panic message including the content of self.

pub fn property_prelude(self) -> Option<CustomPropertyName>

Returns Some if self is of variant PropertyPrelude, and None otherwise.

pub const fn is_counter_style_prelude(&self) -> bool

Returns true if self is of variant CounterStylePrelude.

pub fn as_counter_style_prelude(&self) -> Option<&CustomIdent>

Returns Some if self is a reference of variant CounterStylePrelude, and None otherwise.

pub fn as_mut_counter_style_prelude(&mut self) -> Option<&mut CustomIdent>

Returns Some if self is a mutable reference of variant CounterStylePrelude, and None otherwise.

pub fn expect_counter_style_prelude(self) -> CustomIdent
where Self: Debug,

Unwraps the value, yielding the content of CounterStylePrelude.

Panics

Panics if the value is not CounterStylePrelude, with a panic message including the content of self.

pub fn counter_style_prelude(self) -> Option<CustomIdent>

Returns Some if self is of variant CounterStylePrelude, and None otherwise.

pub const fn is_color_profile_prelude(&self) -> bool

Returns true if self is of variant ColorProfilePrelude.

pub fn as_color_profile_prelude(&self) -> Option<&ColorProfileName>

Returns Some if self is a reference of variant ColorProfilePrelude, and None otherwise.

pub fn as_mut_color_profile_prelude(&mut self) -> Option<&mut ColorProfileName>

Returns Some if self is a mutable reference of variant ColorProfilePrelude, and None otherwise.

pub fn expect_color_profile_prelude(self) -> ColorProfileName
where Self: Debug,

Unwraps the value, yielding the content of ColorProfilePrelude.

Panics

Panics if the value is not ColorProfilePrelude, with a panic message including the content of self.

pub fn color_profile_prelude(self) -> Option<ColorProfileName>

Returns Some if self is of variant ColorProfilePrelude, and None otherwise.

pub const fn is_document_prelude(&self) -> bool

Returns true if self is of variant DocumentPrelude.

pub fn as_document_prelude(&self) -> Option<&DocumentPrelude>

Returns Some if self is a reference of variant DocumentPrelude, and None otherwise.

pub fn as_mut_document_prelude(&mut self) -> Option<&mut DocumentPrelude>

Returns Some if self is a mutable reference of variant DocumentPrelude, and None otherwise.

pub fn expect_document_prelude(self) -> DocumentPrelude
where Self: Debug,

Unwraps the value, yielding the content of DocumentPrelude.

Panics

Panics if the value is not DocumentPrelude, with a panic message including the content of self.

pub fn document_prelude(self) -> Option<DocumentPrelude>

Returns Some if self is of variant DocumentPrelude, and None otherwise.

pub const fn is_font_palette_values_prelude(&self) -> bool

Returns true if self is of variant FontPaletteValuesPrelude.

pub fn as_font_palette_values_prelude(&self) -> Option<&DashedIdent>

Returns Some if self is a reference of variant FontPaletteValuesPrelude, and None otherwise.

pub fn as_mut_font_palette_values_prelude(&mut self) -> Option<&mut DashedIdent>

Returns Some if self is a mutable reference of variant FontPaletteValuesPrelude, and None otherwise.

pub fn expect_font_palette_values_prelude(self) -> DashedIdent
where Self: Debug,

Unwraps the value, yielding the content of FontPaletteValuesPrelude.

Panics

Panics if the value is not FontPaletteValuesPrelude, with a panic message including the content of self.

pub fn font_palette_values_prelude(self) -> Option<DashedIdent>

Returns Some if self is of variant FontPaletteValuesPrelude, and None otherwise.

pub const fn is_font_feature_values_prelude(&self) -> bool

Returns true if self is of variant FontFeatureValuesPrelude.

pub fn as_font_feature_values_prelude( &self ) -> Option<&FontFeatureValuesPrelude>

Returns Some if self is a reference of variant FontFeatureValuesPrelude, and None otherwise.

pub fn as_mut_font_feature_values_prelude( &mut self ) -> Option<&mut FontFeatureValuesPrelude>

Returns Some if self is a mutable reference of variant FontFeatureValuesPrelude, and None otherwise.

pub fn expect_font_feature_values_prelude(self) -> FontFeatureValuesPrelude
where Self: Debug,

Unwraps the value, yielding the content of FontFeatureValuesPrelude.

Panics

Panics if the value is not FontFeatureValuesPrelude, with a panic message including the content of self.

pub fn font_feature_values_prelude(self) -> Option<FontFeatureValuesPrelude>

Returns Some if self is of variant FontFeatureValuesPrelude, and None otherwise.

pub const fn is_nest_prelude(&self) -> bool

Returns true if self is of variant NestPrelude.

pub fn as_nest_prelude(&self) -> Option<&SelectorList>

Returns Some if self is a reference of variant NestPrelude, and None otherwise.

pub fn as_mut_nest_prelude(&mut self) -> Option<&mut SelectorList>

Returns Some if self is a mutable reference of variant NestPrelude, and None otherwise.

pub fn expect_nest_prelude(self) -> SelectorList
where Self: Debug,

Unwraps the value, yielding the content of NestPrelude.

Panics

Panics if the value is not NestPrelude, with a panic message including the content of self.

pub fn nest_prelude(self) -> Option<SelectorList>

Returns Some if self is of variant NestPrelude, and None otherwise.

pub const fn is_keyframes_prelude(&self) -> bool

Returns true if self is of variant KeyframesPrelude.

pub fn as_keyframes_prelude(&self) -> Option<&KeyframesName>

Returns Some if self is a reference of variant KeyframesPrelude, and None otherwise.

pub fn as_mut_keyframes_prelude(&mut self) -> Option<&mut KeyframesName>

Returns Some if self is a mutable reference of variant KeyframesPrelude, and None otherwise.

pub fn expect_keyframes_prelude(self) -> KeyframesName
where Self: Debug,

Unwraps the value, yielding the content of KeyframesPrelude.

Panics

Panics if the value is not KeyframesPrelude, with a panic message including the content of self.

pub fn keyframes_prelude(self) -> Option<KeyframesName>

Returns Some if self is of variant KeyframesPrelude, and None otherwise.

pub const fn is_import_prelude(&self) -> bool

Returns true if self is of variant ImportPrelude.

pub fn as_import_prelude(&self) -> Option<&ImportPrelude>

Returns Some if self is a reference of variant ImportPrelude, and None otherwise.

pub fn as_mut_import_prelude(&mut self) -> Option<&mut ImportPrelude>

Returns Some if self is a mutable reference of variant ImportPrelude, and None otherwise.

pub fn expect_import_prelude(self) -> ImportPrelude
where Self: Debug,

Unwraps the value, yielding the content of ImportPrelude.

Panics

Panics if the value is not ImportPrelude, with a panic message including the content of self.

pub fn import_prelude(self) -> Option<ImportPrelude>

Returns Some if self is of variant ImportPrelude, and None otherwise.

pub const fn is_namespace_prelude(&self) -> bool

Returns true if self is of variant NamespacePrelude.

pub fn as_namespace_prelude(&self) -> Option<&NamespacePrelude>

Returns Some if self is a reference of variant NamespacePrelude, and None otherwise.

pub fn as_mut_namespace_prelude(&mut self) -> Option<&mut NamespacePrelude>

Returns Some if self is a mutable reference of variant NamespacePrelude, and None otherwise.

pub fn expect_namespace_prelude(self) -> NamespacePrelude
where Self: Debug,

Unwraps the value, yielding the content of NamespacePrelude.

Panics

Panics if the value is not NamespacePrelude, with a panic message including the content of self.

pub fn namespace_prelude(self) -> Option<NamespacePrelude>

Returns Some if self is of variant NamespacePrelude, and None otherwise.

pub const fn is_media_prelude(&self) -> bool

Returns true if self is of variant MediaPrelude.

pub fn as_media_prelude(&self) -> Option<&MediaQueryList>

Returns Some if self is a reference of variant MediaPrelude, and None otherwise.

pub fn as_mut_media_prelude(&mut self) -> Option<&mut MediaQueryList>

Returns Some if self is a mutable reference of variant MediaPrelude, and None otherwise.

pub fn expect_media_prelude(self) -> MediaQueryList
where Self: Debug,

Unwraps the value, yielding the content of MediaPrelude.

Panics

Panics if the value is not MediaPrelude, with a panic message including the content of self.

pub fn media_prelude(self) -> Option<MediaQueryList>

Returns Some if self is of variant MediaPrelude, and None otherwise.

pub const fn is_supports_prelude(&self) -> bool

Returns true if self is of variant SupportsPrelude.

pub fn as_supports_prelude(&self) -> Option<&SupportsCondition>

Returns Some if self is a reference of variant SupportsPrelude, and None otherwise.

pub fn as_mut_supports_prelude(&mut self) -> Option<&mut SupportsCondition>

Returns Some if self is a mutable reference of variant SupportsPrelude, and None otherwise.

pub fn expect_supports_prelude(self) -> SupportsCondition
where Self: Debug,

Unwraps the value, yielding the content of SupportsPrelude.

Panics

Panics if the value is not SupportsPrelude, with a panic message including the content of self.

pub fn supports_prelude(self) -> Option<SupportsCondition>

Returns Some if self is of variant SupportsPrelude, and None otherwise.

pub const fn is_page_prelude(&self) -> bool

Returns true if self is of variant PagePrelude.

pub fn as_page_prelude(&self) -> Option<&PageSelectorList>

Returns Some if self is a reference of variant PagePrelude, and None otherwise.

pub fn as_mut_page_prelude(&mut self) -> Option<&mut PageSelectorList>

Returns Some if self is a mutable reference of variant PagePrelude, and None otherwise.

pub fn expect_page_prelude(self) -> PageSelectorList
where Self: Debug,

Unwraps the value, yielding the content of PagePrelude.

Panics

Panics if the value is not PagePrelude, with a panic message including the content of self.

pub fn page_prelude(self) -> Option<PageSelectorList>

Returns Some if self is of variant PagePrelude, and None otherwise.

pub const fn is_layer_prelude(&self) -> bool

Returns true if self is of variant LayerPrelude.

pub fn as_layer_prelude(&self) -> Option<&LayerPrelude>

Returns Some if self is a reference of variant LayerPrelude, and None otherwise.

pub fn as_mut_layer_prelude(&mut self) -> Option<&mut LayerPrelude>

Returns Some if self is a mutable reference of variant LayerPrelude, and None otherwise.

pub fn expect_layer_prelude(self) -> LayerPrelude
where Self: Debug,

Unwraps the value, yielding the content of LayerPrelude.

Panics

Panics if the value is not LayerPrelude, with a panic message including the content of self.

pub fn layer_prelude(self) -> Option<LayerPrelude>

Returns Some if self is of variant LayerPrelude, and None otherwise.

pub const fn is_container_prelude(&self) -> bool

Returns true if self is of variant ContainerPrelude.

pub fn as_container_prelude(&self) -> Option<&ContainerCondition>

Returns Some if self is a reference of variant ContainerPrelude, and None otherwise.

pub fn as_mut_container_prelude(&mut self) -> Option<&mut ContainerCondition>

Returns Some if self is a mutable reference of variant ContainerPrelude, and None otherwise.

pub fn expect_container_prelude(self) -> ContainerCondition
where Self: Debug,

Unwraps the value, yielding the content of ContainerPrelude.

Panics

Panics if the value is not ContainerPrelude, with a panic message including the content of self.

pub fn container_prelude(self) -> Option<ContainerCondition>

Returns Some if self is of variant ContainerPrelude, and None otherwise.

pub const fn is_custom_media_prelude(&self) -> bool

Returns true if self is of variant CustomMediaPrelude.

pub fn as_custom_media_prelude(&self) -> Option<&CustomMediaQuery>

Returns Some if self is a reference of variant CustomMediaPrelude, and None otherwise.

pub fn as_mut_custom_media_prelude(&mut self) -> Option<&mut CustomMediaQuery>

Returns Some if self is a mutable reference of variant CustomMediaPrelude, and None otherwise.

pub fn expect_custom_media_prelude(self) -> CustomMediaQuery
where Self: Debug,

Unwraps the value, yielding the content of CustomMediaPrelude.

Panics

Panics if the value is not CustomMediaPrelude, with a panic message including the content of self.

pub fn custom_media_prelude(self) -> Option<CustomMediaQuery>

Returns Some if self is of variant CustomMediaPrelude, and None otherwise.

pub const fn is_scope_prelude(&self) -> bool

Returns true if self is of variant ScopePrelude.

pub fn as_scope_prelude(&self) -> Option<&ScopeRange>

Returns Some if self is a reference of variant ScopePrelude, and None otherwise.

pub fn as_mut_scope_prelude(&mut self) -> Option<&mut ScopeRange>

Returns Some if self is a mutable reference of variant ScopePrelude, and None otherwise.

pub fn expect_scope_prelude(self) -> ScopeRange
where Self: Debug,

Unwraps the value, yielding the content of ScopePrelude.

Panics

Panics if the value is not ScopePrelude, with a panic message including the content of self.

pub fn scope_prelude(self) -> Option<ScopeRange>

Returns Some if self is of variant ScopePrelude, and None otherwise.