Enum swc_css_ast::ComponentValue

pub enum ComponentValue {

    PreservedToken(Box<TokenAndSpan>),
    Function(Box<Function>),
    SimpleBlock(Box<SimpleBlock>),
    AtRule(Box<AtRule>),
    QualifiedRule(Box<QualifiedRule>),
    ListOfComponentValues(Box<ListOfComponentValues>),
    KeyframeBlock(Box<KeyframeBlock>),
    Ident(Box<Ident>),
    DashedIdent(Box<DashedIdent>),
    Str(Box<Str>),
    Url(Box<Url>),
    Integer(Box<Integer>),
    Number(Box<Number>),
    Percentage(Box<Percentage>),
    Dimension(Box<Dimension>),
    LengthPercentage(Box<LengthPercentage>),
    FrequencyPercentage(Box<FrequencyPercentage>),
    AnglePercentage(Box<AnglePercentage>),
    TimePercentage(Box<TimePercentage>),
    Ratio(Box<Ratio>),
    UnicodeRange(Box<UnicodeRange>),
    Color(Box<Color>),
    AlphaValue(Box<AlphaValue>),
    Hue(Box<Hue>),
    CmykComponent(Box<CmykComponent>),
    Delimiter(Box<Delimiter>),
    CalcSum(Box<CalcSum>),
    ComplexSelector(Box<ComplexSelector>),
    LayerName(Box<LayerName>),
    SupportsCondition(Box<SupportsCondition>),
    Declaration(Box<Declaration>),
    IdSelector(Box<IdSelector>),

}

Variants

PreservedToken(Box<TokenAndSpan>)

Function(Box<Function>)

SimpleBlock(Box<SimpleBlock>)

AtRule(Box<AtRule>)

QualifiedRule(Box<QualifiedRule>)

ListOfComponentValues(Box<ListOfComponentValues>)

KeyframeBlock(Box<KeyframeBlock>)

Ident(Box<Ident>)

DashedIdent(Box<DashedIdent>)

Str(Box<Str>)

Url(Box<Url>)

Integer(Box<Integer>)

Number(Box<Number>)

Percentage(Box<Percentage>)

Dimension(Box<Dimension>)

LengthPercentage(Box<LengthPercentage>)

FrequencyPercentage(Box<FrequencyPercentage>)

AnglePercentage(Box<AnglePercentage>)

TimePercentage(Box<TimePercentage>)

Ratio(Box<Ratio>)

UnicodeRange(Box<UnicodeRange>)

Color(Box<Color>)

AlphaValue(Box<AlphaValue>)

Hue(Box<Hue>)

CmykComponent(Box<CmykComponent>)

Delimiter(Box<Delimiter>)

CalcSum(Box<CalcSum>)

ComplexSelector(Box<ComplexSelector>)

LayerName(Box<LayerName>)

SupportsCondition(Box<SupportsCondition>)

Declaration(Box<Declaration>)

IdSelector(Box<IdSelector>)

Implementations

impl ComponentValue

pub const fn is_preserved_token(&self) -> bool

Returns true if self is of variant PreservedToken.

pub fn as_preserved_token(&self) -> Option<&Box<TokenAndSpan>>

Returns Some if self is a reference of variant PreservedToken, and None otherwise.

pub fn as_mut_preserved_token(&mut self) -> Option<&mut Box<TokenAndSpan>>

Returns Some if self is a mutable reference of variant PreservedToken, and None otherwise.

pub fn expect_preserved_token(self) -> Box<TokenAndSpan>
where Self: Debug,

Unwraps the value, yielding the content of PreservedToken.

Panics

Panics if the value is not PreservedToken, with a panic message including the content of self.

pub fn preserved_token(self) -> Option<Box<TokenAndSpan>>

Returns Some if self is of variant PreservedToken, and None otherwise.

pub const fn is_function(&self) -> bool

Returns true if self is of variant Function.

pub fn as_function(&self) -> Option<&Box<Function>>

Returns Some if self is a reference of variant Function, and None otherwise.

pub fn as_mut_function(&mut self) -> Option<&mut Box<Function>>

Returns Some if self is a mutable reference of variant Function, and None otherwise.

pub fn expect_function(self) -> Box<Function>
where Self: Debug,

Unwraps the value, yielding the content of Function.

Panics

Panics if the value is not Function, with a panic message including the content of self.

pub fn function(self) -> Option<Box<Function>>

Returns Some if self is of variant Function, and None otherwise.

pub const fn is_simple_block(&self) -> bool

Returns true if self is of variant SimpleBlock.

pub fn as_simple_block(&self) -> Option<&Box<SimpleBlock>>

Returns Some if self is a reference of variant SimpleBlock, and None otherwise.

pub fn as_mut_simple_block(&mut self) -> Option<&mut Box<SimpleBlock>>

Returns Some if self is a mutable reference of variant SimpleBlock, and None otherwise.

pub fn expect_simple_block(self) -> Box<SimpleBlock>
where Self: Debug,

Unwraps the value, yielding the content of SimpleBlock.

Panics

Panics if the value is not SimpleBlock, with a panic message including the content of self.

pub fn simple_block(self) -> Option<Box<SimpleBlock>>

Returns Some if self is of variant SimpleBlock, and None otherwise.

pub const fn is_at_rule(&self) -> bool

Returns true if self is of variant AtRule.

pub fn as_at_rule(&self) -> Option<&Box<AtRule>>

Returns Some if self is a reference of variant AtRule, and None otherwise.

pub fn as_mut_at_rule(&mut self) -> Option<&mut Box<AtRule>>

Returns Some if self is a mutable reference of variant AtRule, and None otherwise.

pub fn expect_at_rule(self) -> Box<AtRule>
where Self: Debug,

Unwraps the value, yielding the content of AtRule.

Panics

Panics if the value is not AtRule, with a panic message including the content of self.

pub fn at_rule(self) -> Option<Box<AtRule>>

Returns Some if self is of variant AtRule, and None otherwise.

pub const fn is_qualified_rule(&self) -> bool

Returns true if self is of variant QualifiedRule.

pub fn as_qualified_rule(&self) -> Option<&Box<QualifiedRule>>

Returns Some if self is a reference of variant QualifiedRule, and None otherwise.

pub fn as_mut_qualified_rule(&mut self) -> Option<&mut Box<QualifiedRule>>

Returns Some if self is a mutable reference of variant QualifiedRule, and None otherwise.

pub fn expect_qualified_rule(self) -> Box<QualifiedRule>
where Self: Debug,

Unwraps the value, yielding the content of QualifiedRule.

Panics

Panics if the value is not QualifiedRule, with a panic message including the content of self.

pub fn qualified_rule(self) -> Option<Box<QualifiedRule>>

Returns Some if self is of variant QualifiedRule, and None otherwise.

pub const fn is_list_of_component_values(&self) -> bool

Returns true if self is of variant ListOfComponentValues.

pub fn as_list_of_component_values(&self) -> Option<&Box<ListOfComponentValues>>

Returns Some if self is a reference of variant ListOfComponentValues, and None otherwise.

pub fn as_mut_list_of_component_values( &mut self ) -> Option<&mut Box<ListOfComponentValues>>

Returns Some if self is a mutable reference of variant ListOfComponentValues, and None otherwise.

pub fn expect_list_of_component_values(self) -> Box<ListOfComponentValues>
where Self: Debug,

Unwraps the value, yielding the content of ListOfComponentValues.

Panics

Panics if the value is not ListOfComponentValues, with a panic message including the content of self.

pub fn list_of_component_values(self) -> Option<Box<ListOfComponentValues>>

Returns Some if self is of variant ListOfComponentValues, and None otherwise.

pub const fn is_keyframe_block(&self) -> bool

Returns true if self is of variant KeyframeBlock.

pub fn as_keyframe_block(&self) -> Option<&Box<KeyframeBlock>>

Returns Some if self is a reference of variant KeyframeBlock, and None otherwise.

pub fn as_mut_keyframe_block(&mut self) -> Option<&mut Box<KeyframeBlock>>

Returns Some if self is a mutable reference of variant KeyframeBlock, and None otherwise.

pub fn expect_keyframe_block(self) -> Box<KeyframeBlock>
where Self: Debug,

Unwraps the value, yielding the content of KeyframeBlock.

Panics

Panics if the value is not KeyframeBlock, with a panic message including the content of self.

pub fn keyframe_block(self) -> Option<Box<KeyframeBlock>>

Returns Some if self is of variant KeyframeBlock, and None otherwise.

pub const fn is_ident(&self) -> bool

Returns true if self is of variant Ident.

pub fn as_ident(&self) -> Option<&Box<Ident>>

Returns Some if self is a reference of variant Ident, and None otherwise.

pub fn as_mut_ident(&mut self) -> Option<&mut Box<Ident>>

Returns Some if self is a mutable reference of variant Ident, and None otherwise.

pub fn expect_ident(self) -> Box<Ident>
where Self: Debug,

Unwraps the value, yielding the content of Ident.

Panics

Panics if the value is not Ident, with a panic message including the content of self.

pub fn ident(self) -> Option<Box<Ident>>

Returns Some if self is of variant Ident, and None otherwise.

pub const fn is_dashed_ident(&self) -> bool

Returns true if self is of variant DashedIdent.

pub fn as_dashed_ident(&self) -> Option<&Box<DashedIdent>>

Returns Some if self is a reference of variant DashedIdent, and None otherwise.

pub fn as_mut_dashed_ident(&mut self) -> Option<&mut Box<DashedIdent>>

Returns Some if self is a mutable reference of variant DashedIdent, and None otherwise.

pub fn expect_dashed_ident(self) -> Box<DashedIdent>
where Self: Debug,

Unwraps the value, yielding the content of DashedIdent.

Panics

Panics if the value is not DashedIdent, with a panic message including the content of self.

pub fn dashed_ident(self) -> Option<Box<DashedIdent>>

Returns Some if self is of variant DashedIdent, and None otherwise.

pub const fn is_str(&self) -> bool

Returns true if self is of variant Str.

pub fn as_str(&self) -> Option<&Box<Str>>

Returns Some if self is a reference of variant Str, and None otherwise.

pub fn as_mut_str(&mut self) -> Option<&mut Box<Str>>

Returns Some if self is a mutable reference of variant Str, and None otherwise.

pub fn expect_str(self) -> Box<Str>
where Self: Debug,

Unwraps the value, yielding the content of Str.

Panics

Panics if the value is not Str, with a panic message including the content of self.

pub fn str(self) -> Option<Box<Str>>

Returns Some if self is of variant Str, and None otherwise.

pub const fn is_url(&self) -> bool

Returns true if self is of variant Url.

pub fn as_url(&self) -> Option<&Box<Url>>

Returns Some if self is a reference of variant Url, and None otherwise.

pub fn as_mut_url(&mut self) -> Option<&mut Box<Url>>

Returns Some if self is a mutable reference of variant Url, and None otherwise.

pub fn expect_url(self) -> Box<Url>
where Self: Debug,

Unwraps the value, yielding the content of Url.

Panics

Panics if the value is not Url, with a panic message including the content of self.

pub fn url(self) -> Option<Box<Url>>

Returns Some if self is of variant Url, and None otherwise.

pub const fn is_integer(&self) -> bool

Returns true if self is of variant Integer.

pub fn as_integer(&self) -> Option<&Box<Integer>>

Returns Some if self is a reference of variant Integer, and None otherwise.

pub fn as_mut_integer(&mut self) -> Option<&mut Box<Integer>>

Returns Some if self is a mutable reference of variant Integer, and None otherwise.

pub fn expect_integer(self) -> Box<Integer>
where Self: Debug,

Unwraps the value, yielding the content of Integer.

Panics

Panics if the value is not Integer, with a panic message including the content of self.

pub fn integer(self) -> Option<Box<Integer>>

Returns Some if self is of variant Integer, and None otherwise.

pub const fn is_number(&self) -> bool

Returns true if self is of variant Number.

pub fn as_number(&self) -> Option<&Box<Number>>

Returns Some if self is a reference of variant Number, and None otherwise.

pub fn as_mut_number(&mut self) -> Option<&mut Box<Number>>

Returns Some if self is a mutable reference of variant Number, and None otherwise.

pub fn expect_number(self) -> Box<Number>
where Self: Debug,

Unwraps the value, yielding the content of Number.

Panics

Panics if the value is not Number, with a panic message including the content of self.

pub fn number(self) -> Option<Box<Number>>

Returns Some if self is of variant Number, and None otherwise.

pub const fn is_percentage(&self) -> bool

Returns true if self is of variant Percentage.

pub fn as_percentage(&self) -> Option<&Box<Percentage>>

Returns Some if self is a reference of variant Percentage, and None otherwise.

pub fn as_mut_percentage(&mut self) -> Option<&mut Box<Percentage>>

Returns Some if self is a mutable reference of variant Percentage, and None otherwise.

pub fn expect_percentage(self) -> Box<Percentage>
where Self: Debug,

Unwraps the value, yielding the content of Percentage.

Panics

Panics if the value is not Percentage, with a panic message including the content of self.

pub fn percentage(self) -> Option<Box<Percentage>>

Returns Some if self is of variant Percentage, and None otherwise.

pub const fn is_dimension(&self) -> bool

Returns true if self is of variant Dimension.

pub fn as_dimension(&self) -> Option<&Box<Dimension>>

Returns Some if self is a reference of variant Dimension, and None otherwise.

pub fn as_mut_dimension(&mut self) -> Option<&mut Box<Dimension>>

Returns Some if self is a mutable reference of variant Dimension, and None otherwise.

pub fn expect_dimension(self) -> Box<Dimension>
where Self: Debug,

Unwraps the value, yielding the content of Dimension.

Panics

Panics if the value is not Dimension, with a panic message including the content of self.

pub fn dimension(self) -> Option<Box<Dimension>>

Returns Some if self is of variant Dimension, and None otherwise.

pub const fn is_length_percentage(&self) -> bool

Returns true if self is of variant LengthPercentage.

pub fn as_length_percentage(&self) -> Option<&Box<LengthPercentage>>

Returns Some if self is a reference of variant LengthPercentage, and None otherwise.

pub fn as_mut_length_percentage(&mut self) -> Option<&mut Box<LengthPercentage>>

Returns Some if self is a mutable reference of variant LengthPercentage, and None otherwise.

pub fn expect_length_percentage(self) -> Box<LengthPercentage>
where Self: Debug,

Unwraps the value, yielding the content of LengthPercentage.

Panics

Panics if the value is not LengthPercentage, with a panic message including the content of self.

pub fn length_percentage(self) -> Option<Box<LengthPercentage>>

Returns Some if self is of variant LengthPercentage, and None otherwise.

pub const fn is_frequency_percentage(&self) -> bool

Returns true if self is of variant FrequencyPercentage.

pub fn as_frequency_percentage(&self) -> Option<&Box<FrequencyPercentage>>

Returns Some if self is a reference of variant FrequencyPercentage, and None otherwise.

pub fn as_mut_frequency_percentage( &mut self ) -> Option<&mut Box<FrequencyPercentage>>

Returns Some if self is a mutable reference of variant FrequencyPercentage, and None otherwise.

pub fn expect_frequency_percentage(self) -> Box<FrequencyPercentage>
where Self: Debug,

Unwraps the value, yielding the content of FrequencyPercentage.

Panics

Panics if the value is not FrequencyPercentage, with a panic message including the content of self.

pub fn frequency_percentage(self) -> Option<Box<FrequencyPercentage>>

Returns Some if self is of variant FrequencyPercentage, and None otherwise.

pub const fn is_angle_percentage(&self) -> bool

Returns true if self is of variant AnglePercentage.

pub fn as_angle_percentage(&self) -> Option<&Box<AnglePercentage>>

Returns Some if self is a reference of variant AnglePercentage, and None otherwise.

pub fn as_mut_angle_percentage(&mut self) -> Option<&mut Box<AnglePercentage>>

Returns Some if self is a mutable reference of variant AnglePercentage, and None otherwise.

pub fn expect_angle_percentage(self) -> Box<AnglePercentage>
where Self: Debug,

Unwraps the value, yielding the content of AnglePercentage.

Panics

Panics if the value is not AnglePercentage, with a panic message including the content of self.

pub fn angle_percentage(self) -> Option<Box<AnglePercentage>>

Returns Some if self is of variant AnglePercentage, and None otherwise.

pub const fn is_time_percentage(&self) -> bool

Returns true if self is of variant TimePercentage.

pub fn as_time_percentage(&self) -> Option<&Box<TimePercentage>>

Returns Some if self is a reference of variant TimePercentage, and None otherwise.

pub fn as_mut_time_percentage(&mut self) -> Option<&mut Box<TimePercentage>>

Returns Some if self is a mutable reference of variant TimePercentage, and None otherwise.

pub fn expect_time_percentage(self) -> Box<TimePercentage>
where Self: Debug,

Unwraps the value, yielding the content of TimePercentage.

Panics

Panics if the value is not TimePercentage, with a panic message including the content of self.

pub fn time_percentage(self) -> Option<Box<TimePercentage>>

Returns Some if self is of variant TimePercentage, and None otherwise.

pub const fn is_ratio(&self) -> bool

Returns true if self is of variant Ratio.

pub fn as_ratio(&self) -> Option<&Box<Ratio>>

Returns Some if self is a reference of variant Ratio, and None otherwise.

pub fn as_mut_ratio(&mut self) -> Option<&mut Box<Ratio>>

Returns Some if self is a mutable reference of variant Ratio, and None otherwise.

pub fn expect_ratio(self) -> Box<Ratio>
where Self: Debug,

Unwraps the value, yielding the content of Ratio.

Panics

Panics if the value is not Ratio, with a panic message including the content of self.

pub fn ratio(self) -> Option<Box<Ratio>>

Returns Some if self is of variant Ratio, and None otherwise.

pub const fn is_unicode_range(&self) -> bool

Returns true if self is of variant UnicodeRange.

pub fn as_unicode_range(&self) -> Option<&Box<UnicodeRange>>

Returns Some if self is a reference of variant UnicodeRange, and None otherwise.

pub fn as_mut_unicode_range(&mut self) -> Option<&mut Box<UnicodeRange>>

Returns Some if self is a mutable reference of variant UnicodeRange, and None otherwise.

pub fn expect_unicode_range(self) -> Box<UnicodeRange>
where Self: Debug,

Unwraps the value, yielding the content of UnicodeRange.

Panics

Panics if the value is not UnicodeRange, with a panic message including the content of self.

pub fn unicode_range(self) -> Option<Box<UnicodeRange>>

Returns Some if self is of variant UnicodeRange, and None otherwise.

pub const fn is_color(&self) -> bool

Returns true if self is of variant Color.

pub fn as_color(&self) -> Option<&Box<Color>>

Returns Some if self is a reference of variant Color, and None otherwise.

pub fn as_mut_color(&mut self) -> Option<&mut Box<Color>>

Returns Some if self is a mutable reference of variant Color, and None otherwise.

pub fn expect_color(self) -> Box<Color>
where Self: Debug,

Unwraps the value, yielding the content of Color.

Panics

Panics if the value is not Color, with a panic message including the content of self.

pub fn color(self) -> Option<Box<Color>>

Returns Some if self is of variant Color, and None otherwise.

pub const fn is_alpha_value(&self) -> bool

Returns true if self is of variant AlphaValue.

pub fn as_alpha_value(&self) -> Option<&Box<AlphaValue>>

Returns Some if self is a reference of variant AlphaValue, and None otherwise.

pub fn as_mut_alpha_value(&mut self) -> Option<&mut Box<AlphaValue>>

Returns Some if self is a mutable reference of variant AlphaValue, and None otherwise.

pub fn expect_alpha_value(self) -> Box<AlphaValue>
where Self: Debug,

Unwraps the value, yielding the content of AlphaValue.

Panics

Panics if the value is not AlphaValue, with a panic message including the content of self.

pub fn alpha_value(self) -> Option<Box<AlphaValue>>

Returns Some if self is of variant AlphaValue, and None otherwise.

pub const fn is_hue(&self) -> bool

Returns true if self is of variant Hue.

pub fn as_hue(&self) -> Option<&Box<Hue>>

Returns Some if self is a reference of variant Hue, and None otherwise.

pub fn as_mut_hue(&mut self) -> Option<&mut Box<Hue>>

Returns Some if self is a mutable reference of variant Hue, and None otherwise.

pub fn expect_hue(self) -> Box<Hue>
where Self: Debug,

Unwraps the value, yielding the content of Hue.

Panics

Panics if the value is not Hue, with a panic message including the content of self.

pub fn hue(self) -> Option<Box<Hue>>

Returns Some if self is of variant Hue, and None otherwise.

pub const fn is_cmyk_component(&self) -> bool

Returns true if self is of variant CmykComponent.

pub fn as_cmyk_component(&self) -> Option<&Box<CmykComponent>>

Returns Some if self is a reference of variant CmykComponent, and None otherwise.

pub fn as_mut_cmyk_component(&mut self) -> Option<&mut Box<CmykComponent>>

Returns Some if self is a mutable reference of variant CmykComponent, and None otherwise.

pub fn expect_cmyk_component(self) -> Box<CmykComponent>
where Self: Debug,

Unwraps the value, yielding the content of CmykComponent.

Panics

Panics if the value is not CmykComponent, with a panic message including the content of self.

pub fn cmyk_component(self) -> Option<Box<CmykComponent>>

Returns Some if self is of variant CmykComponent, and None otherwise.

pub const fn is_delimiter(&self) -> bool

Returns true if self is of variant Delimiter.

pub fn as_delimiter(&self) -> Option<&Box<Delimiter>>

Returns Some if self is a reference of variant Delimiter, and None otherwise.

pub fn as_mut_delimiter(&mut self) -> Option<&mut Box<Delimiter>>

Returns Some if self is a mutable reference of variant Delimiter, and None otherwise.

pub fn expect_delimiter(self) -> Box<Delimiter>
where Self: Debug,

Unwraps the value, yielding the content of Delimiter.

Panics

Panics if the value is not Delimiter, with a panic message including the content of self.

pub fn delimiter(self) -> Option<Box<Delimiter>>

Returns Some if self is of variant Delimiter, and None otherwise.

pub const fn is_calc_sum(&self) -> bool

Returns true if self is of variant CalcSum.

pub fn as_calc_sum(&self) -> Option<&Box<CalcSum>>

Returns Some if self is a reference of variant CalcSum, and None otherwise.

pub fn as_mut_calc_sum(&mut self) -> Option<&mut Box<CalcSum>>

Returns Some if self is a mutable reference of variant CalcSum, and None otherwise.

pub fn expect_calc_sum(self) -> Box<CalcSum>
where Self: Debug,

Unwraps the value, yielding the content of CalcSum.

Panics

Panics if the value is not CalcSum, with a panic message including the content of self.

pub fn calc_sum(self) -> Option<Box<CalcSum>>

Returns Some if self is of variant CalcSum, and None otherwise.

pub const fn is_complex_selector(&self) -> bool

Returns true if self is of variant ComplexSelector.

pub fn as_complex_selector(&self) -> Option<&Box<ComplexSelector>>

Returns Some if self is a reference of variant ComplexSelector, and None otherwise.

pub fn as_mut_complex_selector(&mut self) -> Option<&mut Box<ComplexSelector>>

Returns Some if self is a mutable reference of variant ComplexSelector, and None otherwise.

pub fn expect_complex_selector(self) -> Box<ComplexSelector>
where Self: Debug,

Unwraps the value, yielding the content of ComplexSelector.

Panics

Panics if the value is not ComplexSelector, with a panic message including the content of self.

pub fn complex_selector(self) -> Option<Box<ComplexSelector>>

Returns Some if self is of variant ComplexSelector, and None otherwise.

pub const fn is_layer_name(&self) -> bool

Returns true if self is of variant LayerName.

pub fn as_layer_name(&self) -> Option<&Box<LayerName>>

Returns Some if self is a reference of variant LayerName, and None otherwise.

pub fn as_mut_layer_name(&mut self) -> Option<&mut Box<LayerName>>

Returns Some if self is a mutable reference of variant LayerName, and None otherwise.

pub fn expect_layer_name(self) -> Box<LayerName>
where Self: Debug,

Unwraps the value, yielding the content of LayerName.

Panics

Panics if the value is not LayerName, with a panic message including the content of self.

pub fn layer_name(self) -> Option<Box<LayerName>>

Returns Some if self is of variant LayerName, and None otherwise.

pub const fn is_supports_condition(&self) -> bool

Returns true if self is of variant SupportsCondition.

pub fn as_supports_condition(&self) -> Option<&Box<SupportsCondition>>

Returns Some if self is a reference of variant SupportsCondition, and None otherwise.

pub fn as_mut_supports_condition( &mut self ) -> Option<&mut Box<SupportsCondition>>

Returns Some if self is a mutable reference of variant SupportsCondition, and None otherwise.

pub fn expect_supports_condition(self) -> Box<SupportsCondition>
where Self: Debug,

Unwraps the value, yielding the content of SupportsCondition.

Panics

Panics if the value is not SupportsCondition, with a panic message including the content of self.

pub fn supports_condition(self) -> Option<Box<SupportsCondition>>

Returns Some if self is of variant SupportsCondition, and None otherwise.

pub const fn is_declaration(&self) -> bool

Returns true if self is of variant Declaration.

pub fn as_declaration(&self) -> Option<&Box<Declaration>>

Returns Some if self is a reference of variant Declaration, and None otherwise.

pub fn as_mut_declaration(&mut self) -> Option<&mut Box<Declaration>>

Returns Some if self is a mutable reference of variant Declaration, and None otherwise.

pub fn expect_declaration(self) -> Box<Declaration>
where Self: Debug,

Unwraps the value, yielding the content of Declaration.

Panics

Panics if the value is not Declaration, with a panic message including the content of self.

pub fn declaration(self) -> Option<Box<Declaration>>

Returns Some if self is of variant Declaration, and None otherwise.

pub const fn is_id_selector(&self) -> bool

Returns true if self is of variant IdSelector.

pub fn as_id_selector(&self) -> Option<&Box<IdSelector>>

Returns Some if self is a reference of variant IdSelector, and None otherwise.

pub fn as_mut_id_selector(&mut self) -> Option<&mut Box<IdSelector>>

Returns Some if self is a mutable reference of variant IdSelector, and None otherwise.

pub fn expect_id_selector(self) -> Box<IdSelector>
where Self: Debug,

Unwraps the value, yielding the content of IdSelector.

Panics

Panics if the value is not IdSelector, with a panic message including the content of self.

pub fn id_selector(self) -> Option<Box<IdSelector>>

Returns Some if self is of variant IdSelector, and None otherwise.