Macro swc_css_ast::matches_eq

macro_rules! matches_eq {
    ($value:expr, $($pat:expr),*) => { ... };
}

Returns true if the given value matches one of the given patterns.

The type of value and patterns should be identical.

Examples

use swc_atoms::Atom;
use swc_atoms::js_word;
use swc_css_ast::*;

assert!(matches_eq!(Atom::from("a"), "a"));
assert!(matches_eq!("a", "a"));