Crate swc_css_ast

AST definitions for CSS.

Macros

matches_eq

Returns true if the given value matches one of the given patterns.

matches_eq_ignore_ascii_case

Returns true if the given value matches one of the given patterns.

Structs

AnPlusBNotation

Angle

AnyNamespace

AtRule

AttributeSelector

AttributeSelectorMatcher

AttributeSelectorModifier

CalcOperator

CalcProduct

CalcSum

ClassSelector

Combinator

ComplexSelector

CompoundSelector

e.g. foo.c1.c2

CompoundSelectorList

ContainerCondition

ContainerQuery

ContainerQueryAnd

ContainerQueryNot

ContainerQueryOr

CustomHighlightName

CustomIdent

CustomMediaQuery

CustomPropertyName

DashedIdent

Declaration

Delimiter

DimensionToken

DocumentPrelude

ExtensionName

Flex

FontFeatureValuesPrelude

ForgivingRelativeSelectorList

ForgivingSelectorList

Frequency

Function

HexColor

IdSelector

Ident

ImportConditions

ImportPrelude

ImportantFlag

Integer

KeyframeBlock

KeyframesPseudoFunction

KeyframesPseudoPrefix

LayerName

LayerNameList

Length

ListOfComponentValues

MediaAnd

MediaCondition

MediaConditionWithoutOr

MediaFeatureBoolean

MediaFeaturePlain

MediaFeatureRange

MediaFeatureRangeInterval

MediaNot

MediaOr

MediaQuery

MediaQueryList

NamedNamespace

NamespacePrefix

NamespacePrelude

NestingSelector

Number

PageSelector

PageSelectorList

PageSelectorPseudo

PageSelectorType

Percentage

PseudoClassSelector

PseudoElementSelector

QualifiedRule

Ratio

RelativeSelector

RelativeSelectorList

Resolution

ScopeRange

SelectorList

SequenceOfCustomIdents

SimpleBlock

SizeFeatureBoolean

SizeFeaturePlain

SizeFeatureRange

SizeFeatureRangeInterval

Str

Quoted string.

Stylesheet

SupportsAnd

SupportsCondition

SupportsNot

SupportsOr

TagNameSelector

Time

TokenAndSpan

UnicodeRange

UniversalSelector

UnknownDimension

Url

UrlKeyValue

UrlValueRaw

WqName

Enums

AbsoluteColorBase

AlphaValue

AnPlusB

AnglePercentage

AtRuleName

AtRulePrelude

AttributeSelectorMatcherValue

AttributeSelectorValue

BinOp

CalcOperatorType

CalcProductOrOperator

CalcValue

CalcValueOrOperator

CmykComponent

Color

ColorProfileName

CombinatorValue

ComplexSelectorChildren

ComponentValue

ContainerName

ContainerQueryType

CustomMediaQueryMediaType

DeclarationName

DeclarationOrAtRule

DelimiterValue

Dimension

DocumentPreludeMatchingFunction

FamilyName

ForgivingComplexSelector

ForgivingRelativeSelector

FrequencyPercentage

FunctionName

GeneralEnclosed

Hue

ImportHref

ImportLayerName

KeyframeSelector

KeyframesName

LayerPrelude

LengthPercentage

MediaConditionAllType

MediaConditionType

MediaConditionWithoutOrType

MediaFeature

MediaFeatureName

MediaFeatureRangeComparison

MediaFeatureValue

MediaInParens

MediaType

Namespace

NamespacePreludeUri

NumberType

PseudoClassSelectorChildren

PseudoElementSelectorChildren

QualifiedRulePrelude

QueryInParens

Rule

SizeFeature

SizeFeatureName

SizeFeatureRangeComparison

SizeFeatureValue

StyleBlock

SubclassSelector

SupportsConditionType

SupportsFeature

SupportsInParens

TimePercentage

Token

TypeSelector

UrlModifier

UrlValue